Robert Corral Code Review
It is not hard to follow what's going on.. but I believe the point of commenting now is to practice expressing what something does so something more complicated is easier to explain. Whether or not you can do that I have no idea. if I read the code, your comment is sparse and leaves me wanting to know more information about how things are being done. other than that I have been learning by a suggestion about making"Pythonic" Code. Google it and try :D